requirements.txt
# Trans-border Representation Probe - Python Dependencies
# =======================================================
# Installation: pip install -r requirements.txt

# ============================================================
# Core API & HTTP
# ============================================================
openai>=1.0.0                    # OpenAI-compatible SDK for OpenRouter API
requests>=2.31.0                 # HTTP library for API calls
tenacity>=8.2.0                  # Retry logic with exponential backoff

# ============================================================
# Data Processing & Analysis
# ============================================================
pandas>=1.5.0                    # Data manipulation and analysis
numpy>=1.24.0                    # Numerical computing

# ============================================================
# Embedding & NLP
# ============================================================
sentence-transformers>=2.2.0     # Multilingual embedding models
scikit-learn>=1.3.0              # Machine learning utilities (cosine similarity, t-SNE)
scipy>=1.10.0                    # Scientific computing (correlation analysis)

# ============================================================
# Visualization
# ============================================================
matplotlib>=3.7.0                # Plotting library
seaborn>=0.12.0                  # Statistical data visualization

# ============================================================
# Configuration & Environment
# ============================================================
python-dotenv>=1.0.0             # Environment variable management (.env file support)
pyyaml>=6.0                      # YAML configuration file parsing

# ============================================================
# Development & Testing (Optional)
# ============================================================
pytest>=7.4.0                    # Testing framework
pytest-cov>=4.1.0                # Test coverage reporting
black>=23.0.0                    # Code formatter
flake8>=6.0.0                    # Linting
mypy>=1.5.0                      # Type checking

# ============================================================
# Jupyter Notebook Support (Optional)
# ============================================================
jupyter>=1.0.0                   # Jupyter notebook
ipykernel>=6.25.0                # IPython kernel for Jupyter
ipywidgets>=8.0.0                # Interactive widgets

# ============================================================
# Additional Utilities
# ============================================================
tqdm>=4.65.0                     # Progress bars
openpyxl>=3.1.0                  # Excel file support for pandas
